add late assist test

+++ system/sensord/rawgps/test_rawgps.py
@@ -24,6 +24,11 @@ class TestRawgpsd(unittest.TestCase):
 
     cls.sm = messaging.SubMaster(['qcomGnss', 'gpsLocation', 'gnssMeasurements'])
 
+
+  def setUp(self):
+    at_cmd("AT+QGPSDEL=0")
+
+
   def tearDown(self):
     managed_processes['rawgpsd'].stop()
     os.system("sudo systemctl restart systemd-resolved")
@@ -42,20 +47,11 @@ class TestRawgpsd(unittest.TestCase):
     self._wait_for_output(30)
     assert self.sm.updated['qcomGnss']
 
-
-  def test_startup_time_no_internet(self):
-    os.system("sudo systemctl stop systemd-resolved")
-    for _ in range(1):
+  def test_startup_time(self):
+    for i in range(2):
+      if i == 1:
+        os.system("sudo systemctl stop systemd-resolved")
       managed_processes['rawgpsd'].start()
-
-      self._wait_for_output(7)
-      assert self.sm.updated['qcomGnss']
-      managed_processes['rawgpsd'].stop()
-
-  def test_startup_time_internet(self):
-    for _ in range(1):
-      managed_processes['rawgpsd'].start()
-
       self._wait_for_output(7)
       assert self.sm.updated['qcomGnss']
       managed_processes['rawgpsd'].stop()
@@ -70,43 +66,49 @@ class TestRawgpsd(unittest.TestCase):
       loc_status = json.loads(ls)
       assert set(loc_status['modem']['location']['enabled']) <= {'3gpp-lac-ci'}
 
-  def test_assistance_loading(self):
-    # clear assistance data
-    at_cmd("AT+QGPSDEL=0")
 
-    managed_processes['rawgpsd'].start()
-    self._wait_for_output(10)
-    assert self.sm.updated['qcomGnss']
-    managed_processes['rawgpsd'].stop()
-
+  def check_assistance(self, should_be_loaded):
     # after QGPSDEL: '+QGPSXTRADATA: 0,"1980/01/05,19:00:00"'
     # after loading: '+QGPSXTRADATA: 10080,"2023/06/24,19:00:00"'
     out = at_cmd("AT+QGPSXTRADATA?")
     out = out.split("+QGPSXTRADATA:")[1].split("'")[0].strip()
     valid_duration, injected_time_str = out.split(",", 1)
-    assert valid_duration == "10080"  # should be max time
-    injected_time = datetime.datetime.strptime(injected_time_str.replace("\"", ""), "%Y/%m/%d,%H:%M:%S")
-    self.assertLess(abs((datetime.datetime.utcnow() - injected_time).total_seconds()), 60*60*12)
+    if should_be_loaded:
+      assert valid_duration == "10080"  # should be max time
+      injected_time = datetime.datetime.strptime(injected_time_str.replace("\"", ""), "%Y/%m/%d,%H:%M:%S")
+      self.assertLess(abs((datetime.datetime.utcnow() - injected_time).total_seconds()), 60*60*12)
+    else:
+      valid_duration, injected_time_str = out.split(",", 1)
+      injected_time_str = injected_time_str.replace('\"', '').replace('\'', '')
+      assert injected_time_str[:] == '1980/01/05,19:00:00'[:]
+      assert valid_duration == '0'
+
+  def test_assistance_loading(self):
+    managed_processes['rawgpsd'].start()
+    self._wait_for_output(10)
+    assert self.sm.updated['qcomGnss']
+    managed_processes['rawgpsd'].stop()
+    self.check_assistance(True)
 
   def test_no_assistance_loading(self):
     os.system("sudo systemctl stop systemd-resolved")
-    # clear assistance data
-    at_cmd("AT+QGPSDEL=0")
 
     managed_processes['rawgpsd'].start()
     self._wait_for_output(10)
     assert self.sm.updated['qcomGnss']
     managed_processes['rawgpsd'].stop()
+    self.check_assistance(False)
 
-    # after QGPSDEL: '+QGPSXTRADATA: 0,"1980/01/05,19:00:00"'
-    # after loading: '+QGPSXTRADATA: 10080,"2023/06/24,19:00:00"'
-    out = at_cmd("AT+QGPSXTRADATA?")
-    out = out.split("+QGPSXTRADATA:")[1].split("'")[0].strip()
-    valid_duration, injected_time_str = out.split(",", 1)
-    injected_time_str = injected_time_str.replace('\"', '').replace('\'', '')
-    assert injected_time_str[:] == '1980/01/05,19:00:00'[:]
-    assert valid_duration == '0'
+  def test_late_assistance_loading(self):
+    os.system("sudo systemctl stop systemd-resolved")
 
+    managed_processes['rawgpsd'].start()
+    self._wait_for_output(15)
+    assert self.sm.updated['qcomGnss']
+    os.system("sudo systemctl restart systemd-resolved")
+    self._wait_for_output(15)
+    managed_processes['rawgpsd'].stop()
+    self.check_assistance(True)
 
   @unittest.skipIf(not GOOD_SIGNAL, "No good GPS signal")
   def test_fix(self):
